First cut at pulling a selection of the C implementation of rules (plus a Makefile) into this new environment
I've cut the number of files down compared with the initial
implementation of this stuff in my private 'rulesc' git repo. Trying
to avoid having to hack on header files that aren't located in the
same directory as the other rules-related C files, to make the
integration with "new-day" repo smoother.

Anyway, we'll see how this flies...